+++ app/src/main/java/com/boostcamp/and03/ui/screen/canvasmemo/CanvasMemoScreen.kt
@@ -1,13 +1,10 @@
 package com.boostcamp.and03.ui.screen.canvasmemo
 
 import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
-import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.gestures.detectTransformGestures
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.WindowInsetsSides
@@ -37,19 +34,11 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Path
 import androidx.compose.ui.graphics.TransformOrigin
-import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.drawText
-import androidx.compose.ui.text.rememberTextMeasurer
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
@@ -64,6 +53,8 @@ import com.boostcamp.and03.ui.screen.canvasmemo.component.AddQuoteBottomSheet
 import com.boostcamp.and03.ui.screen.canvasmemo.component.AddQuoteDialog
 import com.boostcamp.and03.ui.screen.canvasmemo.component.AlertAction
 import com.boostcamp.and03.ui.screen.canvasmemo.component.AlertMessageCard
+import com.boostcamp.and03.ui.screen.canvasmemo.component.Arrows
+import com.boostcamp.and03.ui.screen.canvasmemo.component.DraggableCanvasItem
 import com.boostcamp.and03.ui.screen.canvasmemo.component.NodeItem
 import com.boostcamp.and03.ui.screen.canvasmemo.component.QuoteItem
 import com.boostcamp.and03.ui.screen.canvasmemo.component.RelationEditorDialog
@@ -71,7 +62,6 @@ import com.boostcamp.and03.ui.screen.canvasmemo.component.ToolAction
 import com.boostcamp.and03.ui.screen.canvasmemo.component.ToolExpandableButton
 import com.boostcamp.and03.ui.screen.canvasmemo.component.bottombar.MainBottomBar
 import com.boostcamp.and03.ui.screen.canvasmemo.component.bottombar.MainBottomBarType
-import com.boostcamp.and03.ui.screen.canvasmemo.model.EdgeUiModel
 import com.boostcamp.and03.ui.screen.canvasmemo.model.MemoNodeUiModel
 import com.boostcamp.and03.ui.screen.canvasmemo.model.RelationAddStep
 import com.boostcamp.and03.ui.theme.And03Padding
@@ -254,7 +244,6 @@ private fun CanvasMemoScreen(
                                 }
                             }
                         }
-
                 ) {
                     Box(
                         modifier = Modifier
@@ -581,126 +570,6 @@ private fun CanvasMemoScreen(
                             }
                         )
                     }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun DraggableCanvasItem(
-    nodeId: String,
-    worldOffset: Offset,
-    onMove: (Offset) -> Unit,
-    onSizeChanged: (IntSize) -> Unit,
-    modifier: Modifier = Modifier,
-    content: @Composable BoxScope.() -> Unit,
-    onClick: ((String) -> Unit)? = null,
-    draggable: Boolean = true,
-) {
-    Box(
-        modifier = modifier
-            .graphicsLayer {
-                translationX = worldOffset.x
-                translationY = worldOffset.y
-            }
-            .onGloballyPositioned { coords ->
-                onSizeChanged(coords.size)
-            }
-            .then(
-                if (onClick != null) {
-                    Modifier.pointerInput(nodeId) {
-                        detectTapGestures(
-                            onTap = {
-                                onClick(nodeId)
-                            }
-                        )
-                    }
-                } else {
-                    Modifier
-                }
-            )
-            .then(
-                if (draggable) {
-                    Modifier.pointerInput(nodeId) {
-                        detectDragGestures { change, dragAmount ->
-                            change.consume()
-                            onMove(dragAmount)
-                        }
-                    }
-                } else Modifier
-            )
-    ) {
-        content()
-    }
-}
-
-@Composable
-fun Arrows(
-    arrows: List<EdgeUiModel>,
-    items: Map<String, MemoNodeUiModel>,
-    nodeSizes: Map<String, IntSize>,
-) {
-    val textMeasurer = rememberTextMeasurer()
-
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        arrows.forEach { edge ->
-
-            val fromNode = items[edge.edge.fromId]
-            val toNode = items[edge.edge.toId]
-            val fromSize = nodeSizes[edge.edge.fromId] ?: IntSize.Zero
-            val toSize = nodeSizes[edge.edge.toId] ?: IntSize.Zero
-
-            if (fromNode != null && toNode != null) {
-
-                val start = fromNode.node.offset + Offset(
-                    fromSize.width.toFloat(),
-                    fromSize.height / 2f
-                )
-
-                val end = toNode.node.offset + Offset(
-                    0f,
-                    toSize.height / 2f
-                )
-
-                val midX = (start.x + end.x) / 2
-
-                val path = Path().apply {
-                    moveTo(start.x, start.y)
-                    lineTo(midX, start.y)
-                    lineTo(midX, end.y)
-                    lineTo(end.x, end.y)
-                }
-
-                drawPath(
-                    path = path,
-                    color = Color.Black,
-                    style = Stroke(width = 4f)
-                )
-
-                val label = edge.edge.name
-
-                if (label.isNotEmpty()) {
-                    val textLayoutResult = textMeasurer.measure(label)
-                    val textWidth = textLayoutResult.size.width
-                    val textHeight = textLayoutResult.size.height
-
-                    val textPos = Offset(
-                        x = midX - textWidth / 2,
-                        y = (start.y + end.y) / 2 - textHeight / 2
-                    )
-
-                    drawRect(
-                        color = Color.White,
-                        topLeft = textPos,
-                        size = Size(textWidth.toFloat(), textHeight.toFloat())
-                    )
-
-                    // 텍스트 그리기
-                    drawText(
-                        textLayoutResult = textLayoutResult,
-                        topLeft = textPos
-                    )
                 }
             }
         }
