+++ app/src/main/java/com/boostcamp/and03/ui/screen/canvasmemo/component/DraggableCanvasItem.kt
@@ -1,0 +1,61 @@
+package com.boostcamp.and03.ui.screen.canvasmemo.component
+
+import androidx.compose.foundation.gestures.detectDragGestures
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.onGloballyPositioned
+import androidx.compose.ui.unit.IntSize
+
+@Composable
+fun DraggableCanvasItem(
+    nodeId: String,
+    worldOffset: Offset,
+    onMove: (Offset) -> Unit,
+    onSizeChanged: (IntSize) -> Unit,
+    modifier: Modifier = Modifier,
+    content: @Composable BoxScope.() -> Unit,
+    onClick: ((String) -> Unit)? = null,
+    draggable: Boolean = true,
+) {
+    Box(
+        modifier = modifier
+            .graphicsLayer {
+                translationX = worldOffset.x
+                translationY = worldOffset.y
+            }
+            .onGloballyPositioned { coords ->
+                onSizeChanged(coords.size)
+            }
+            .then(
+                if (onClick != null) {
+                    Modifier.pointerInput(nodeId) {
+                        detectTapGestures(
+                            onTap = {
+                                onClick(nodeId)
+                            }
+                        )
+                    }
+                } else {
+                    Modifier
+                }
+            )
+            .then(
+                if (draggable) {
+                    Modifier.pointerInput(nodeId) {
+                        detectDragGestures { change, dragAmount ->
+                            change.consume()
+                            onMove(dragAmount)
+                        }
+                    }
+                } else Modifier
+            )
+    ) {
+        content()
+    }
+}
